add has_runnable and update results roll up
