+++ lib/inferno/test_runner.rb
@@ -171,7 +171,10 @@ module Inferno
 
     def roll_up_result(results)
       result_priority = Entities::Result::RESULT_OPTIONS
-      unique_results = results.map(&:result).uniq
+      unique_required_results = results.select(&:required?)
+      all_optional = unique_required_results.empty?
+      unique_results = all_optional ? results : unique_required_results
+      unique_results = unique_results.map(&:result).uniq
       result_priority.find do |result|
         unique_results.include? result
       end
